FlaskServer/requirement.txt
Flask
tensorflow
numpy
Pillow
librosa
matplotlib